brymen-dmm: Support BM857s detection
Per the protocol document linked from https://sigrok.org/wiki/Brymen_BM857
RTS must be toggled on connection to wake up the device.  At least on a
brand new BM859s, this was indeed required.

Unfortunately, this approach relies on directly reaching into the
libserialport API from the driver, so is not upstreamable in this form.

Signed-off-by: Karl Palsson <karlp@tweak.net.au>