korad-kaxxxxp: speed up scan process
The Korad protocool relies on unterminated strings, which is workable
for the commands, but the variable length replies of the IDN response
make it hard.  We know from history that some devices can take a long
time to reply, so we must have some decent timeout/retry, but we also
shouldn't keep waiting when we know we're not likely to receive any
data. To this end, we keep the long retry loop, but only until we start
receiving data.  Once the device has started replying, we don't expect
any futher delays other than due to the baud rate.

This speeds up scan from ~5 seconds to ~0.1 seconds on a KA3005P v5.5

Signed-off-by: Karl Palsson <karlp@tweak.net.au>